+++ migrations/000001_create_user_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE users (
     login varchar(255) NOT NULL,
     password varchar(255) NOT NULL,
-    balance double precision DEFAULT 0.0,
+    balance numeric(30,2) DEFAULT 0.0,
     created_at timestamp NOT NULL,
     PRIMARY KEY(login)
 )
